refactor(startup): isolate app-wide composition from MaterialApp
_DivineAppState.build() built the whole app in one expression: the
MultiRepositoryProvider / MultiBlocProvider / global-listener tree, the
MaterialApp at its leaf, and UploadFailureListener's implementation below it
in the same file. Nothing in that tree could be mounted, inspected or tested
without the entrypoint.

Split it three ways:

- AppCompositionRoot — every app-wide repository, bloc and global listener,
  with the app injected as `child`. The locale and appearance cubits it
  provides are what the MaterialApp reads, so the nesting direction is now
  explicit in the type instead of implied by a closure.
- DivineMaterialApp — the MaterialApp.router, reading those two cubits
  itself rather than through two nested BlocBuilders and a captured builder
  closure.
- UploadFailureListener moves to its own file with the five private helpers
  that only it used, so main.dart no longer carries an app-wide UI listener.

build() is now the deep-link listener, the Android back-button channel
registration, the crash probe, and a three-line composition.

One consequence worth knowing: the ref.watch calls in the composition tree
used to rebuild all of _DivineAppState, re-running BackButtonHandler.initialize
each time. They now rebuild only AppCompositionRoot.

main.dart: 3450 -> 2224 lines, 154 -> 103 imports.
